+++ sigpwny.com/src/components/Meeting/Row.astro
@@ -5,8 +5,8 @@ import { getCollection, getEntries } from 'astro:content';
 import { PdfSvg, YouTubeSvg } from '@/components/Icons';
 import { CountdownBadge } from '@/components/ReactMigration/Countdown';
 import { TagGroup } from '@/components/ReactMigration/Tag';
-import { convertDate, weekNumber } from '@/lib/meetings';
-import { getProfilesFromNames } from '@/lib/profiles';
+import { convertDate, weekNumber } from '@/utils/meetings';
+import { getProfilesFromNames } from '@/utils/profiles';
 
 import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
