+++ sigpwny.com/src/pages/meetings/[...slug]/index.astro
@@ -19,8 +19,8 @@ import {
   getMeetings,
   getYouTubeEmbedUrl,
   weekNumber
-} from '@/lib/meetings';
-import { getProfilesFromNames } from '@/lib/profiles';
+} from '@/utils/meetings';
+import { getProfilesFromNames } from '@/utils/profiles';
 // import { Tooltip } from '@/components/Tooltip';
 
 import dayjs from 'dayjs';
