+++ sigpwny.com/src/pages/meetings/index.astro
@@ -11,7 +11,7 @@ import {
   getMeetings,
   formatSemester,
   weekNumber
-} from "@/lib/meetings";
+} from "@/utils/meetings";
 
 const rawMeetings = (await getMeetings()).sort(
   (a, b) => b.data.time_start.valueOf() - a.data.time_start.valueOf()
